Further robustness improvements and clean-up. - Changes to commands.h functions to return the devctl() return code - Changes to commands.h function read_frame_raw_noblock() to skip error reporting for return code EAGAIN, which is normal operation, not error; i.e. "There are no messages in the queue." - Improvements to driver-flood-tests.cpp to make it more robust. - Further verification checks added to driver-micro-tests.cpp. - Added '-Wall' compiler option to CMakeLists.txt to align with QNX Momentic Makefile and associated minor fixes to address warning. - Fixed some string hexadecimal reading issues in tool cansend. - Minor fixes to log messages.

+++ tests/timer/timer-tests.cpp
@@ -40,7 +40,7 @@ extern "C" {
 
         test_data->trigger_time_ms = now;
 
-        for (int i = 0; i < test_data->size; ++i) {
+        for (size_t i = 0; i < test_data->size; ++i) {
             test_data->sum += test_data->numbers[i];
         }
     }
